frontend/src/app/shared/components/shared/key-result-overview/key-result-overview.component.ts
import { Component, OnInit } from '@angular/core';
import { ActivatedRoute } from '@angular/router';
import { Observable, switchMap } from 'rxjs';
import { Goal, GoalService } from '../../../services/goal.service';
import { Location } from '@angular/common';
import { getNumberOrNull } from '../../../common';

@Component({
  selector: 'app-key-result-overview',
  templateUrl: './key-result-overview.component.html',
  styleUrls: ['./key-result-overview.component.scss'],
})
export class KeyResultOverviewComponent implements OnInit {
  public goal$!: Observable<Goal>;

  constructor(
    private goalService: GoalService,
    private route: ActivatedRoute,
    private location: Location
  ) {}

  ngOnInit(): void {
    this.goal$ = this.route.paramMap.pipe(
      switchMap((params) => {
        const keyResultId = getNumberOrNull(params.get('keyresultId'));
        if (keyResultId) {
          return this.goalService.getGoalByKeyResultId(keyResultId);
        } else {
          throw Error('KeyResult with Id ' + keyResultId + " doesn't exist");
        }
      })
    );
  }

  navigateBack() {
    this.location.back();
  }
}
